move functions to where they are used
